Updated component configuration, need a way to remove devices when reconfiguring or restarting

+++ custom_components/pihole_v6/models/summary.py
@@ -87,4 +87,4 @@ class PiHoleSummary(BaseModel):
     queries: Queries
     clients: Clients
     gravity: Gravity
-    took: int
+    took: float
